LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/update-the-image-data/current-request-microtask-002-expected.txt



FAIL Current request microtask handling with multiple tasks. assert_equals: currentSrc should be  expected "" but got "http://localhost:8800/images/green.png?0.566006149334529"
